+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # Changelog
 
+## 0.9.2 (x)
+* #19 Create a tool library for jupytab, allowing separated download for jupytab util
+* #7 Jupytab must be install in kernel as no-deps
+* #16 notebook.js use parent path instead of same level path
+
+## 0.9.1 (2019-12-23)
+* #3 Multiindex column dataframe fails in Tableau
+* #8 Jupytab not working on windows
+* #10 Improve Table column name management
+* #11 Add option to export also index from dataframe
+
 ## 0.9.0 (2019-06-30)
 * Jupytab public release
