+++ jupytab-utils/jupytab_utils/__init__.py
@@ -1,6 +1,7 @@
 # Copyright (c) 2019 Capital Fund Management
 # SPDX-License-Identifier: MIT
 
-from .util import Tables, DataFrameTable
+from .table import Tables
+from .dataframe_table import DataFrameTable
 
 __all__ = [Tables, DataFrameTable]
