+++ jupytab-utils/requirements-dev.txt
@@ -1,5 +1,3 @@
-ipykernel
-pandas
 pytest
 flake8
 pytest-cov
