+++ jupytab-utils/requirements.txt
@@ -1,0 +1,1 @@
+pandas<1
