+++ jupytab-utils/setup.cfg
