+++ jupytab-utils/tests/test_util.py
@@ -4,22 +4,21 @@
 import os
 
 import pandas as pd
-
-from jupytab import Tables, DataFrameTable
+import jupytab_utils as jtu
 
 THIS_DIR = os.path.abspath(os.path.dirname(__file__))
-RESOURCES = os.path.join(THIS_DIR, '../samples/real-estate_crime')
+RESOURCES = os.path.join(THIS_DIR, 'resources')
 
 
 def test_data_schema():
     crime_df = pd.read_csv(os.path.join(RESOURCES, 'sacramento_crime.csv'))
     realestate_df = pd.read_csv(os.path.join(RESOURCES, 'sacramento_realestate.csv'))
 
-    tables = Tables()
+    tables = jtu.Tables()
     tables['sacramento_crime'] = \
-        DataFrameTable('Sacramento Crime', dataframe=crime_df)
+        jtu.DataFrameTable('Sacramento Crime', dataframe=crime_df)
     tables['sacramento_realestate'] = \
-        DataFrameTable('Sacramento RealEstate', dataframe=realestate_df)
+        jtu.DataFrameTable('Sacramento RealEstate', dataframe=realestate_df)
 
     schema = tables.schema()
 
@@ -44,9 +43,9 @@ dataType": "int"}, {"id": "latitude", "dataType": "float"}, {"id": "longitude", 
 
 
 def test_clean_column_name():
-    assert DataFrameTable.clean_column_name(['abéçpo$ù"', 0, 'AaZz_#"\\']) == "_abecpo_u_0_AaZz__"
+    assert jtu.DataFrameTable.clean_column_name(['abéçpo$ù"', 0, 'AaZz_#"\\']) == "_abecpo_u_0_AaZz__"
 
 
 def test_replace_duplicated_column_name():
-    assert DataFrameTable.replace_duplicated_column_name(['A', 'A', 'a', 'z', 'a', 'Y']) \
+    assert jtu.DataFrameTable.replace_duplicated_column_name(['A', 'A', 'a', 'z', 'a', 'Y']) \
            == ['A_1', 'A_2', 'a_1', 'z', 'a_2', 'Y']
