+++ jupytab/MANIFEST.in
