+++ jupytab/docs/resources/ConfigSection.png
