+++ jupytab/docs/resources/RealEstateCrime.png
