+++ jupytab/docs/resources/component-overview.png
