+++ jupytab/jupytab/log_pipe.py
