+++ jupytab/jupytab/static/favicon.ico
