+++ jupytab/jupytab/static/notebook.js
