+++ jupytab/requirements-dev.txt
@@ -1,0 +1,5 @@
+ipykernel
+pytest
+flake8
+pytest-cov
+coverage
