--- requirements.txt
@@ -1,1 +1,0 @@
-jupyter_kernel_gateway
